+++ CHANGES.md
@@ -8,7 +8,8 @@ Warning: this release introduces some changes which break the backward compatibi
 * The OSGi Bundle-SymbolicName is changed from `javax.util.streamex` to `one.util.streamex`.
 * The Maven groupID is changed from `io.github.amaembo` to `one.util`. Dependencies in pom.xml files should be updated accordingly.
 * Added: `StreamEx.ofLines(Path)` and `StreamEx.ofLines(Path, Charset)`
-* Added: `MoreCollectors.commonPrefix()/commonSuffix()` short-circuiting collectors. 
+* Added: `MoreCollectors.commonPrefix()/commonSuffix()` short-circuiting collectors.
+* Added: `IntStreamEx.of(Integer[])`, `LongStreamEx.of(Long[])`, `DoubleStreamEx.of(Double[])` static methods.   
 * Deprecated methods `StreamEx.ofEntries()` removed.
 * Deprecated methods `collectingAndThen` in primitive collectors removed (use `andThen()` instead).